Ensure all files are correctly published.
Unfortunately, npm uses `.gitignore` by default, as config of which
files to deploy and which to not. Anything listed in the
`.gitignore` will not be published. We obviously have `*.js` in our
`.gitignore` as we write `.es6` code.

Adding a `files` directive to our package.json tells npm to
specifically include particular files, overriding the contents of
`.gitignore`.